tipo de atributo money por integer

+++ database/migrations/2017_07_20_020416_create_coffees_table.php
@@ -15,7 +15,7 @@ class CreateCoffeesTable extends Migration
         Schema::create('coffees', function (Blueprint $table) {
             $table->increments('id');
             $table->string('nombre');
-            $table->money('precio');
+            $table->integer('precio');
             $table->timestamps();
         });
     }
